5/07/2020 daniele

+++ app/src/main/java/it/giudevo/worldbank/ViewListContents.java
@@ -4,9 +4,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.database.Cursor;
 import android.graphics.Color;
-import android.graphics.Typeface;
 import android.os.Bundle;
-import android.os.Parcelable;
 import android.util.Log;
 import android.util.SparseBooleanArray;
 import android.view.Menu;
@@ -14,7 +12,6 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
-import android.widget.ListAdapter;
 import android.widget.ListView;
 import android.widget.Toast;
 
@@ -27,8 +24,9 @@ public class ViewListContents extends AppCompatActivity implements AdapterView.O
     ListView lvFav;
     public Cursor res;
     public Cursor data;
-    private SparseBooleanArray selectedList = new SparseBooleanArray();
+    public SparseBooleanArray selectedList = new SparseBooleanArray();
     public ArrayList<String> theList;
+    public ArrayAdapter<String> listAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -49,9 +47,27 @@ public class ViewListContents extends AppCompatActivity implements AdapterView.O
                 theList.add(data.getString(1));
                 Log.w("CA", data.getString(1));
 
-                ListAdapter listAdapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, theList);
+                Log.w("CA", "RESTORE ADAPTER");
+                listAdapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, theList);
                 lvFav.setAdapter(listAdapter);
-                lvFav.setOnItemLongClickListener(this);
+                lvFav.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
+                    @Override
+                    public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
+
+                        boolean isSelected = selectedList.get(position,false);
+                        if(isSelected) {
+                            parent.setSelected(true);
+                            selectedList.delete(position);
+                            parent.getChildAt(position).setBackgroundColor(Color.WHITE);
+                        } else {
+                            parent.setSelected(false);
+                            selectedList.put(position, true);
+                            parent.getChildAt(position).setBackgroundColor(Color.GRAY);
+                        }
+                        Log.w("CA", String.valueOf(selectedList.size()));
+                        return true;
+                    }
+                });
                 lvFav.setOnItemClickListener(new AdapterView.OnItemClickListener() {
                     @Override
                     public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
@@ -65,38 +81,40 @@ public class ViewListContents extends AppCompatActivity implements AdapterView.O
                 });
 
             }
+            Log.w("CA", "la lista e" + theList);
         }
     }
 
     @Override
     public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
 
-//        boolean isSelected = selectedList.get(position,false);
-//        if(isSelected) {
-//            parent.setSelected(false);
-//            //setTypeface(null, Typeface.BOLD);
-//        //  parent.getChildAt(position).setBackgroundColor(Color.WHITE);
-//        } else {
-//            parent.setSelected(true);
-//            //setTypeface(null, Typeface.NORMAL);
-//            parent.getChildAt(position).setBackgroundColor(Color.GRAY);
-//        }
 
-        boolean isSel = theList.get(position).isEmpty();
-
-
-        if (isSel) {
-            parent.setSelected(false);
-            //selectedList.delete(position);
+        boolean isSelected = selectedList.get(position,false);
+        if(isSelected) {
+            parent.setSelected(true);
+            selectedList.delete(position);
             parent.getChildAt(position).setBackgroundColor(Color.WHITE);
         } else {
-            parent.setSelected(true);
+            parent.setSelected(false);
             selectedList.put(position, true);
-            parent.getChildAt(position).setBackgroundColor(Color.GREEN);
-            String idd = theList.get(position);
-            myDB.deletedata(idd);
-            theList.remove(position);
+            parent.getChildAt(position).setBackgroundColor(Color.GRAY);
         }
+        Log.w("CA", String.valueOf(selectedList.size()));
+
+//        boolean isSel = theList.get(position).isEmpty();
+//
+//        if (isSel) {
+//            parent.setSelected(false);
+//            selectedList.delete(position);/////////////////////////////////////////
+//            parent.getChildAt(position).setBackgroundColor(Color.RED);
+//        } else {
+//            parent.setSelected(true);
+//            selectedList.put(position, true);
+//            parent.getChildAt(position).setBackgroundColor(Color.GREEN);
+//            //String idd = theList.get(position);
+//            //myDB.deletedata(idd);
+//            //theList.remove(position);
+//        }
         return true;
 
     }
@@ -111,31 +129,53 @@ public class ViewListContents extends AppCompatActivity implements AdapterView.O
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
-
                 case R.id.delete_all:{
                     Log.w("accesso a  del", "granted");
                     for (int index = theList.size() - 1; index >= 0; index--) {
-                    String id = theList.get(index);
-                    myDB.deletedata(id);
-                    theList.remove(index);
-                   }
-                selectedList.clear();
+                        if(selectedList.get(index)) {
+                            String id = theList.get(index);
+                            myDB.deletedata(id);
+                            theList.remove(index);
+                            //theList.clear();
+
+                        }
+                    }
+                    selectedList.clear();
+                    listAdapter.notifyDataSetChanged();
+                    Log.w("CA", "DELETE");
                 break;
                 }
 
 
                 case R.id.unselect_all:{
-                selectedList.clear();
-                break;}
+                    for(int i = 0; i < theList.size()-1; i++){
+                        //theList.set(i);
+                        selectedList.delete(i);
+                        //theList.set(i);
+                        //theList.clear();
+
+                    }
+                    selectedList.clear();
+                    listAdapter.notifyDataSetChanged();
+                    break;
+                }
 
 
                 case R.id.select_all:{
-                for (int i = 0; i < theList.size(); i++) {
-                    selectedList.put(i, true);
+                    for (int i = 0; i < theList.size(); i++) {
+                        //Toast.makeText(this, selectedList.size(), Toast.LENGTH_LONG).show();
+                        Log.w("CA", selectedList.size() + "aaaaaaaaaaaaaaaaaaaaaaaaaa");
+                        selectedList.put(i, true);
+                    }
+                    selectedList.clear();
+                    listAdapter.notifyDataSetChanged();
+                    //listAdapter.notifyDataSetInvalidated();
+                    //listAdapter.setNotifyOnChange(true);
+                    break;
                 }
-                break;}
 
         }
         return true;
     }
+
 }
